+++ realm/realm-annotations-processor/src/test/java/io/realm/processor/RealmProcessorTest.java
@@ -59,14 +59,6 @@ public class RealmProcessorTest {
     private JavaFileObject UseExtendRealmList = JavaFileObjects.forResource("some/test/UseExtendRealmList.java");
     private JavaFileObject SimpleRealmModel = JavaFileObjects.forResource("some/test/SimpleRealmModel.java");
     private JavaFileObject customInterface = JavaFileObjects.forResource("some/test/CustomInterface.java");
-    private JavaFileObject backlinks = JavaFileObjects.forResource("some/test/Backlinks.java");
-    private JavaFileObject backlinksTarget = JavaFileObjects.forResource("some/test/BacklinkTarget.java");
-    private JavaFileObject counter = JavaFileObjects.forResource("some/test/Counter.java");
-    private JavaFileObject counterIgnored = JavaFileObjects.forResource("some/test/Counter_Ignored.java");
-    private JavaFileObject counterIndexed = JavaFileObjects.forResource("some/test/Counter_Indexed.java");
-    private JavaFileObject counterPK = JavaFileObjects.forResource("some/test/Counter_PK.java");
-    private JavaFileObject counterRequired = JavaFileObjects.forResource("some/test/Counter_Required.java");
-    private JavaFileObject counterStatic = JavaFileObjects.forResource("some/test/Counter_Static.java");
     private JavaFileObject nonLatinName = JavaFileObjects.forResource("some/test/ÁrvíztűrőTükörfúrógép.java");
 
     @Test
@@ -477,213 +469,10 @@ public class RealmProcessorTest {
     }
 
     @Test
-    public void compileMutableRealmInteger() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(counter))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void compileIgnoredMutableRealmInteger() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(counterIgnored))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void compileIndexedMutableRealmInteger() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(counterIndexed))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void failOnPKMutableRealmInteger() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(counterPK))
-                .processedWith(new RealmProcessor())
-                .failsToCompile();
-    }
-
-    @Test
-    public void compileRequiredMutableRealmInteger() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(counterRequired))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void compileStaticMutableRealmInteger() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(counterStatic))
-                .processedWith(new RealmProcessor())
-                .failsToCompile();
-    }
-
-    @Test
-    public void compileBacklinks() {
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinks, backlinksTarget))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void compileSyntheticBacklinks() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void failOnLinkingObjectsWithInvalidFieldType() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // Backlinks must be RealmResults
-                .type("BacklinkTarget")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("Fields annotated with @LinkingObjects must be RealmResults");
-    }
-
-    @Test
-    public void failOnLinkingObjectsWithNonFinalField() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // A field with a @LinkingObjects annotation must be final
-                .modifiers(Modifier.PUBLIC)
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("must be final");
-    }
-
-    @Test
-    public void failsOnLinkingObjectsWithLinkedFields() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // Defining a backlink more than one levels back is not supported.
-                // It can be queried though: equalTo("selectedFieldParents.selectedFieldParents")
-                .clearAnnotations()
-                .annotation("LinkingObjects(\"child.id\")")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("The use of '.' to specify fields in referenced classes is not supported");
-    }
-
-    @Test
-    public void failsOnLinkingObjectsMissingFieldName() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // No backlinked field specified
-                .clearAnnotations()
-                .annotation("LinkingObjects")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("must have a parameter identifying the link target");
-    }
-
-    @Test
-    public void failsOnLinkingObjectsMissingGeneric() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // No backlink generic param specified
-                .type("RealmResults")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("must specify a generic type");
-    }
-
-    @Test
-    public void failsOnLinkingObjectsWithRequiredFields() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // A backlinked field may not be @Required
-                .annotation("Required")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("cannot be @Required");
-    }
-
-    @Test
-    public void failsOnLinkingObjectsWithIgnoreFields() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // An  @Ignored, backlinked field is completely ignored
-                .annotation("Ignore")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .compilesWithoutError();
-    }
-
-    @Test
-    public void failsOnLinkingObjectsFieldNotFound() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // The argument to the @LinkingObjects annotation must name a field in the target class
-                .clearAnnotations()
-                .annotation("LinkingObjects(\"xxx\")")
-                .builder().build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("does not exist in class");
-    }
-
-    @Test
-    public void failsOnLinkingObjectsWithFieldWrongType() throws IOException {
-        RealmSyntheticTestClass javaFileObject = createBacklinkTestClass()
-                // The type of the field named in the @LinkingObjects annotation must match
-                // the generic type of the annotated field.  BacklinkTarget.child is a Backlink,
-                // not a Backlinks_WrongType.
-                .builder().name("Backlinks_WrongType").build();
-        ASSERT.about(javaSources())
-                .that(Arrays.asList(backlinksTarget, javaFileObject))
-                .processedWith(new RealmProcessor())
-                .failsToCompile()
-                .withErrorContaining("instead of");
-    }
-
-    @Test
     public void compareNonLatinName() {
         ASSERT.about(javaSource())
                 .that(nonLatinName)
                 .processedWith(new RealmProcessor())
                 .compilesWithoutError();
-    }
-
-    // This method constructs a synthetic Backlinks test class that *should* compile correctly.
-    // It returns the ref to the backlinked Field.  Tests can modify the
-    // field in perverse ways, to verify failure modes.
-    private RealmSyntheticTestClass.Field createBacklinkTestClass() {
-        return new RealmSyntheticTestClass.Builder().name("Backlinks")
-                .field().name("id").type("int").builder()
-                .field()
-                    .name("parents")
-                    .type("RealmResults<BacklinkTarget>")
-                    .modifiers(Modifier.PUBLIC, Modifier.FINAL)
-                    .annotation("LinkingObjects(\"child\")")
-                    .initializer("null")
-                    .hasGetter(false)
-                    .hasSetter(false);
     }
 }
