--- realm/realm-annotations-processor/src/test/resources/some/test/Counter_Ignored.java
@@ -1,16 +1,0 @@
-package some.test;
-
-import io.realm.MutableRealmInteger;
-import io.realm.RealmObject;
-import io.realm.annotations.Ignore;
-
-public class Counter_Ignored extends RealmObject {
-    private int id;
-
-    @Ignore
-    private final MutableRealmInteger columnMutableRealmInteger = MutableRealmInteger.valueOf(0);
-
-    public MutableRealmInteger getColumnMutableRealmInteger() {
-        return columnMutableRealmInteger;
-    }
-}
