--- realm/realm-annotations-processor/src/test/resources/some/test/Counter_PK.java
@@ -1,15 +1,0 @@
-package some.test;
-
-import io.realm.MutableRealmInteger;
-import io.realm.RealmObject;
-import io.realm.annotations.PrimaryKey;
-
-public class Counter_PK extends RealmObject {
-
-    @PrimaryKey
-    private final MutableRealmInteger columnMutableRealmInteger = MutableRealmInteger.valueOf(0);
-
-    public MutableRealmInteger getColumnMutableRealmInteger() {
-        return columnMutableRealmInteger;
-    }
-}
