--- realm/realm-annotations-processor/src/test/resources/some/test/Counter_Static.java
@@ -1,13 +1,0 @@
-package some.test;
-
-import io.realm.MutableRealmInteger;
-import io.realm.RealmObject;
-
-public class Counter_Static extends RealmObject {
-
-    private static final MutableRealmInteger columnMutableRealmInteger = MutableRealmInteger.valueOf(0);
-
-    public MutableRealmInteger getColumnMutableRealmInteger() {
-        return columnMutableRealmInteger;
-    }
-}
